SPRINT-4587: iOS App crashes when playing a stream that has concurrency limit enabled
I think this fix finally resolves the issues we've had with the ObjectLoader getting deallocated before it tries to send out notifications - resulting in a crash.

A shouldFinalizeLoad flag has been introduced. This makes sure to call the finalizeLoad method, when appropriate, *before* the ObjectLoader gets deallocated.

In other words:
If the order of execution didn't matter, I could just move the calling of informDelegateOfError(which in turn deallocates self) to the end of the containing method. But since I don't want to mess with RestKit internals more than I have to, I preserve the order of execution and just make sure the finalizeLoad method is called when self (the ObjectLoader) still exists, ie. just before it gets deallocated.